Updated file names, login and logout function, started TMDB token all
I updated the file names so they all follow the same naming. Removed the loginCurl.php due to PDO issue being fixed, STARTED the function to be able to add/update 3rd party credentials. Almost got TMDB connection done. updated some session variables.
