+++ Web/Media_tracker_Web/src/php/authentication/login.php
@@ -17,6 +17,7 @@
             if ($result) {
                 session_start();
                 $_SESSION['username'] = $username;
+                $_SESSION['signed_in'] = true;
 
                 // Fetch the user_platform_ids for all platforms
                 $idStmt = $pdo->prepare("SELECT platform_id, user_platform_id FROM useraccounts WHERE username = ?");
